issue #131 refactor: warning modal move to common, feat: handle Delete Siren fuction add
