+++ src/components/Siren/Detail/Detail.tsx
@@ -1,23 +1,38 @@
 import { useParams, useNavigate } from "react-router-dom";
 
 import { Flex, Box, Divider } from "@/components/common";
+import DeleteWarningModal from "@/components/common/WarningModal/DeleteWarningModal/DeleteWarningModal";
 import { Title, Content, Comment } from "@/components/Siren/Detail";
 
 import { useSirenQuery } from "@/hooks/api/useSirenQuery";
+import useModal from "@/hooks/useModal";
 
 import { layoutStyle } from "@/components/Siren/Detail/Detail.style";
 
 const Detail = () => {
 	const param = useParams();
 
-	const { sirenData } = useSirenQuery(Number(param.id));
+	const sirenId = Number(param.id);
+
+	const { sirenData } = useSirenQuery(sirenId);
 
 	const navigate = useNavigate();
+
+	const modal = useModal();
 
 	const handleEditSiren = () => {
 		if (!sirenData) return;
 
-		navigate(`/siren/view/${param.id}?mode=edit`);
+		navigate(`/siren/view/${sirenId}?mode=edit`);
+	};
+
+	const handleDeleteSiren = () => {
+		modal.openModal({
+			key: `DeleteWarningModal`,
+			component: () => <DeleteWarningModal targetId={sirenId} target="siren" />,
+			isUpper: true,
+			notCloseIcon: true,
+		});
 	};
 
 	return (
@@ -31,6 +46,7 @@ const Detail = () => {
 						lostDate={sirenData.result.lostDate}
 						viewCount={sirenData.result.viewCount}
 						handleEditSiren={handleEditSiren}
+						handleDeleteSiren={handleDeleteSiren}
 					/>
 					<Divider />
 					<Content
