+++ src/components/Story/StoryDetail/Comment/Reply/ReplyItem.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 
 import { Flex, Box, Text } from "@/components/common";
-import DeleteWarningModal from "@/components/Story/StoryDetail/DeleteWarningModal/DeleteWarningModal";
+import DeleteWarningModal from "@/components/common/WarningModal/DeleteWarningModal/DeleteWarningModal";
 import Profile from "@/components/Story/StoryDetail/Profile/Profile";
 
 import useModal from "@/hooks/useModal";
