+++ src/components/Story/StoryDetail/StoryDetail.tsx
@@ -3,9 +3,9 @@ import { useState, useEffect, useRef, useCallback } from "react";
 import HeartEmptyIcon from "@/assets/svg/ic-heart-empty.svg?react";
 
 import { Flex, Box, Divider, Text, Carousel } from "@/components/common";
+import DeleteWarningModal from "@/components/common/WarningModal/DeleteWarningModal/DeleteWarningModal";
 import Comment from "@/components/Story/StoryDetail/Comment/Comment";
 import CommentInput from "@/components/Story/StoryDetail/Comment/CommentInput";
-import DeleteWarningModal from "@/components/Story/StoryDetail/DeleteWarningModal/DeleteWarningModal";
 import Profile from "@/components/Story/StoryDetail/Profile/Profile";
 import StoryEdit from "@/components/Story/StoryUpload/StoryEdit/StoryEdit";
